Implement a BVH
Implements a bounded volume hierarchy acceleration structure. It's
supposed to massively improve ray intersection checking performance.

However, with the current amount of objects in the scene the resulting
performance is much worse than before, likely due to the BVH traversal
adding a non-trivial amount of overhead (or I've just done some dumb
mistake somewhere). Might remove this later.

I decreased the number of samples to 10 to make the frame rate
comparable to before.